Se agregaron validaciones de los cruds de productos y categorias

+++ src/category/category.controller.js
@@ -24,6 +24,10 @@ export const saveCategory = (request,response)=>{
 export const updateCategory = async (request,response)=>{
     try {
         let {id_category} = request.params
+        let isValidaIdCategory = await Category.findOne({_id:id_category})
+        if(!isValidaIdCategory){
+            return response.status(400).send({sucess:false,message:'Category not found'})
+        }
         let data = request.body
         let updatedCategory = await Category.findByIdAndUpdate(id_category,data,{new:true})
         response.status(200).send({sucess:true,message:'Category updated succesfully',updatedCategory})
@@ -35,6 +39,10 @@ export const updateCategory = async (request,response)=>{
 export const deleteCategory = async(request,response)=>{
     try {
         let {id_category} = request.params
+        let isValidaIdCategory = await Category.findOne({_id:id_category})
+        if(!isValidaIdCategory){
+            return response.status(400).send({sucess:false,message:'Category not found'})
+        }
         let deletedCategory = await Category.findByIdAndDelete(id_category)
         response.status(200).send({sucess:true,message:'Category deleted succesfully',deletedCategory})
     } catch (error) {
