+++ src/products/product.controller.js
@@ -1,4 +1,3 @@
-import { response } from 'express'
 import Category from '../category/category.model.js'
 import Product from '../products/product.model.js'
 
@@ -14,9 +13,10 @@ export const getProducts = async(request,response)=>{
 export const saveProducts =async (request,response)=>{
     try {
         let data = request.body
-        console.log(data);
-        
         let isValidCategory = await Category.findOne({_id:data.category})
+        if(!isValidCategory){
+            return response.status(400).send({success:false,message:'Category Id not found'})
+        }
         let newProduct = new Product(data)
         await newProduct.save()
         response.status(200).send({success:true,message:'Product saved succesfully'})
@@ -29,9 +29,17 @@ export const updateProducts = async(request,response)=>{
     try {
         let {id_product} = request.params
         let data = request.body
+        let isValidProduct = await Product.findOne({_id:id_product})
+        if(!isValidProduct){
+            return response.status(400).send({success:false,message:'Product Id not found'})
+        }
+        let isValidCategory = await Category.findOne({_id:data.category})
+        if(!isValidCategory){
+            return response.status(400).send({success:false,message:'Category Id not found'})
+        }
         let updateProduct = await Product.findByIdAndUpdate(id_product,data,{new:true})
         console.log(updateProduct);
-        response.status(200).send({success:true,message:'Product Updated Succesfully'})
+        response.status(200).send({success:true,message:'Product Updated Succesfully',updateProduct})
     } catch (error) {
         response.status(500).send({success:false,message:'General Server error',error})
     }
@@ -40,6 +48,10 @@ export const updateProducts = async(request,response)=>{
 export const deleteProduct = async(request,response)=>{
     try {
         let {id_product} = request.params
+        let isValidProduct = await Product.findOne({_id:id_product})
+        if(!isValidProduct){
+            return response.status(400).send({success:false,message:'Product Id not found'})
+        }
         let productDeleted = await Product.findByIdAndDelete(id_product)
         response.status(200).send({success:true,message:'Product Deleted Succesfully',productDeleted})
     } catch (error) {
